Further changes to Basic Properties + Level of Robustness
(1) Put all the properties on top. Looks much better that way.
(2) Put all the discussion right afterwards.
(3) Put the proof at the end, as it seems to take away from the
discussion
(4) Put the twice differentiability around 1 with phi''(1)>0 condition
for Pardo's result.
(5) Expanded the Basic Properties discussion a bit