Section 4.3(Additional on Suppress) ++
(1) Changed running title to be the same as the title of the paper
(2) Added some more notation (s_w) and technical small details after the
dual problem is introduced in the Formulation Section. Reworded there a
bit.
(3) Rewrote ssec:suppress to look more polished & understandable
(4) This resulted in some changes to the ssec:pop